work towards cleaning
