+++ src/PEPSKit.jl
@@ -17,8 +17,10 @@ include("utility/diffset.jl")
 include("utility/hook_pullback.jl")
 include("utility/autoopt.jl")
 
+
 include("states/abstractpeps.jl")
 include("states/infinitepeps.jl")
+
 
 include("operators/transferpeps.jl")
 include("operators/infinitepepo.jl")
@@ -27,15 +29,25 @@ include("operators/derivatives.jl")
 include("operators/localoperator.jl")
 include("operators/models.jl")
 
-include("environments/ctmrg environments.jl")
+
+include("environments/ctmrg_environments.jl")
 include("environments/transferpeps_environments.jl")
 include("environments/transferpepo_environments.jl")
 
+
+include("algorithms/contractions/localoperator.jl")
+
 include("algorithms/ctmrg/ctmrg.jl")
 include("algorithms/ctmrg/gaugefix.jl")
-include("algorithms/ctmrg/ctmrg_all_sides.jl")
+
+include("algorithms/toolbox.jl")
 
 include("algorithms/peps_opt.jl")
+
+
+
+# include("algorithms/ctmrg/ctmrg_all_sides.jl")
+
 
 include("utility/symmetrization.jl")
