+++ src/algorithms/toolbox.jl
@@ -1,0 +1,96 @@
+function MPSKit.expectation_value(peps::InfinitePEPS, O::LocalOperator, envs::CTMRGEnv)
+    checklattice(peps, O)
+    return sum(O.terms) do (inds, operator)
+        contract_localoperator(inds, operator, peps, peps, envs) /
+        contract_localnorm(inds, peps, peps, envs)
+    end
+end
+
+function costfun(peps::InfinitePEPS, envs::CTMRGEnv, O::LocalOperator)
+    E = MPSKit.expectation_value(peps, O, envs)
+    ignore_derivatives() do
+        isapprox(imag(E), 0; atol=sqrt(eps(real(E)))) ||
+            @warn "Expectation value is not real: $E."
+    end
+    return real(E)
+end
+
+function LinearAlgebra.norm(peps::InfinitePEPS, env::CTMRGEnv)
+    total = one(scalartype(peps))
+
+    for r in 1:size(peps, 1), c in 1:size(peps, 2)
+        rprev = _prev(r, size(peps, 1))
+        rnext = _next(r, size(peps, 1))
+        cprev = _prev(c, size(peps, 2))
+        cnext = _next(c, size(peps, 2))
+        total *= @autoopt @tensor env.edges[WEST, r, cprev][χ1 D1 D2; χ2] *
+            env.corners[NORTHWEST, rprev, cprev][χ2; χ3] *
+            env.edges[NORTH, rprev, c][χ3 D3 D4; χ4] *
+            env.corners[NORTHEAST, rprev, cnext][χ4; χ5] *
+            env.edges[EAST, r, cnext][χ5 D5 D6; χ6] *
+            env.corners[SOUTHEAST, rnext, cnext][χ6; χ7] *
+            env.edges[SOUTH, rnext, c][χ7 D7 D8; χ8] *
+            env.corners[SOUTHWEST, rnext, cprev][χ8; χ1] *
+            peps[r, c][d; D3 D5 D7 D1] *
+            conj(peps[r, c][d; D4 D6 D8 D2])
+        total *= tr(
+            env.corners[NORTHWEST, rprev, cprev] *
+            env.corners[NORTHEAST, rprev, c] *
+            env.corners[SOUTHEAST, r, c] *
+            env.corners[SOUTHWEST, r, cprev],
+        )
+        total /= @autoopt @tensor env.edges[WEST, r, cprev][χ1 D1 D2; χ2] *
+            env.corners[NORTHWEST, rprev, cprev][χ2; χ3] *
+            env.corners[NORTHEAST, rprev, c][χ3; χ4] *
+            env.edges[EAST, r, c][χ4 D1 D2; χ5] *
+            env.corners[SOUTHEAST, rnext, c][χ5; χ6] *
+            env.corners[SOUTHWEST, rnext, cprev][χ6; χ1]
+        total /= @autoopt @tensor env.corners[NORTHWEST, rprev, cprev][χ1; χ2] *
+            env.edges[NORTH, rprev, c][χ2 D1 D2; χ3] *
+            env.corners[NORTHEAST, rprev, cnext][χ3; χ4] *
+            env.corners[SOUTHEAST, r, cnext][χ4; χ5] *
+            env.edges[SOUTH, r, c][χ5 D1 D2; χ6] *
+            env.corners[SOUTHWEST, r, cprev][χ6; χ1]
+    end
+
+    return total
+end
+
+"""
+    correlation_length(peps::InfinitePEPS, env::CTMRGEnv; howmany=2)
+
+Compute the PEPS correlation length based on the horizontal and vertical
+transfer matrices. Additionally the (normalized) eigenvalue spectrum is
+returned. Specify the number of computed eigenvalues with `howmany`.
+"""
+function MPSKit.correlation_length(peps::InfinitePEPS, env::CTMRGEnv; num_vals=2)
+    T = scalartype(peps)
+    ξ_h = Vector{real(T)}(undef, size(peps, 1))
+    ξ_v = Vector{real(T)}(undef, size(peps, 2))
+    λ_h = Vector{Vector{T}}(undef, size(peps, 1))
+    λ_v = Vector{Vector{T}}(undef, size(peps, 2))
+
+    # Horizontal
+    above_h = MPSMultiline(map(r -> InfiniteMPS(env.edges[1, r, :]), 1:size(peps, 1)))
+    respaced_edges_h = map(zip(space.(env.edges)[1, :, :], env.edges[3, :, :])) do (V1, T3)
+        return TensorMap(T3.data, V1)
+    end
+    below_h = MPSMultiline(map(r -> InfiniteMPS(respaced_edges_h[r, :]), 1:size(peps, 1)))
+    transfer_peps_h = TransferPEPSMultiline(peps, NORTH)
+    vals_h = MPSKit.transfer_spectrum(above_h, transfer_peps_h, below_h; num_vals)
+    λ_h = map(λ_row -> λ_row / abs(λ_row[1]), vals_h)  # Normalize largest eigenvalue
+    ξ_h = map(λ_row -> -1 / log(abs(λ_row[2])), λ_h)
+
+    # Vertical
+    above_v = MPSMultiline(map(c -> InfiniteMPS(env.edges[2, :, c]), 1:size(peps, 2)))
+    respaced_edges_v = map(zip(space.(env.edges)[2, :, :], env.edges[4, :, :])) do (V2, T4)
+        return TensorMap(T4.data, V2)
+    end
+    below_v = MPSMultiline(map(c -> InfiniteMPS(respaced_edges_v[:, c]), 1:size(peps, 2)))
+    transfer_peps_v = TransferPEPSMultiline(peps, EAST)
+    vals_v = MPSKit.transfer_spectrum(above_v, transfer_peps_v, below_v; num_vals)
+    λ_v = map(λ_row -> λ_row / abs(λ_row[1]), vals_v)  # Normalize largest eigenvalue
+    ξ_v = map(λ_row -> -1 / log(abs(λ_row[2])), λ_v)
+
+    return ξ_h, ξ_v, λ_h, λ_v
+end
